Fix the fundamental problem with the game
This is a doozy LOL. I looked long and hard trying to figure out why
one press of the 3 buttons suddenly made the numbers go so high.

My first impression was the I needed a way to break out of the loop,
but I couldn't use the keyword "break" since event listeners aren't
compatible.

Then, I realized the problem isn't breaking the loop, it's why one
button is causing the scores to go so high in the first place! Then I
realized, it's the loop!

I no longer need the for loop because the loop was causing one event
to turn into MANY events which caused the score to go so high.
Do I know necessarily why those numbers occured? No. But, the
fundamental problem of the game is fixed and it is bare bones playable
up to 5 rounds!